i32 rather than usize

+++ src/one.rs
@@ -1,14 +1,14 @@
-pub fn first(input: String) -> usize {
+pub fn first(input: String) -> i32 {
     let input = preprocess(input);
     count_increased(1, input)
 }
 
-pub fn second(input: String) -> usize {
+pub fn second(input: String) -> i32 {
     let input = preprocess(input);
     count_increased(3, input)
 }
 
-fn count_increased(width: usize, data: Vec<i32>) -> usize {
+fn count_increased(width: usize, data: Vec<i32>) -> i32 {
     data.windows(width)
         .map(|w| -> i32 { w.iter().sum() })
         .fold((0, None), |(count, last), w| match last {
